Trying to use some JavaScript
The code isn't working 100%, but I wanted to make a commit just so I have a saved version of this code that I can refer to later. I think I found a way to fix it (which I will test and commit in a bit) but I wanted to commit this first just in case the new way doesn't work or if there's a better way to do it.