Loading JSON (not fully functional)
Just wanted to make a commit so I have an updated version on Github. I think my issue is probably due to a variable being defined as the wrong type (possibly because the file I'm loading might be a list of JSON objects as opposed to just a single JSON object?) or I could be assigning it to the wrong variable?

I also added a giant image of a puppy (hosted on Firebase) to my index page. I'll probably change it to a pug soon!