Fixed UsingJSON.html!!
The articles now display properly! It turns out that the variable "theData" in UsingJSON.html was a different type than what I thought.

I think I'm almost done with the RSS feed too, but I'm having a bit of trouble converting the loaded RSS to the right format (i.e. converting RSS or XML to JSON.)